changed the type

+++ include/sense_eeg.h
@@ -17,7 +17,7 @@ public:
     ~GtecEEGPublisher();
     int num_channels;
     int num_samples;
-    float sampling_rate;
+    int sampling_rate;
     std::string serial_num;
     rclcpp::Publisher<eeg_msgs::msg::EEGBlock>::SharedPtr publisher;
